A library was created to which classes: Cards, Deck and CardDeckFiller were transferred.
