+++ CardsDeck.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31424.327
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CardsDeck", "CardsDeck\CardsDeck.csproj", "{313C1A58-1C27-4AC3-A2EE-DA9863DEFA6E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CardsDeck", "CardsDeck\CardsDeck.csproj", "{313C1A58-1C27-4AC3-A2EE-DA9863DEFA6E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyLibrary", "MyLibrary\MyLibrary.csproj", "{42C7EA8E-F513-4F67-B3AA-D13EC8DA91DF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{313C1A58-1C27-4AC3-A2EE-DA9863DEFA6E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{313C1A58-1C27-4AC3-A2EE-DA9863DEFA6E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{313C1A58-1C27-4AC3-A2EE-DA9863DEFA6E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{42C7EA8E-F513-4F67-B3AA-D13EC8DA91DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{42C7EA8E-F513-4F67-B3AA-D13EC8DA91DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{42C7EA8E-F513-4F67-B3AA-D13EC8DA91DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{42C7EA8E-F513-4F67-B3AA-D13EC8DA91DF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
